% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/AMAPVox.R
\docType{package}
\name{AMAPVox}
\alias{AMAPVox}
\alias{AMAPVox-package}
\title{AMAPVox package}
\description{
The package provides a a set of R functions for working with voxel spaces
(read, write, plot, etc.).
Voxel spaces are read from text-based output files of the
\url{https://amap-dev.cirad.fr/projects/amapvox}{AMAPVox software}.
}
\section{References}{
 \itemize{ \item{ Research paper first describing
AMAPVox:\cr Vincent, G., Antin, C., Laurans, M., Heurtebize, J., Durrieu,
S., Lavalley, C., & Dauzat, J. (2017). Mapping plant area index of tropical
evergreen forest by airborne laser scanning. A cross-validation study using
LAI2200 optical sensor. Remote Sensing of Environment, 198, 254-266.
\doi{10.1016/j.rse.2017.05.034} } \item{ Up-to-date description of PAD/LAD
estimators implemented in AMAPVox:\cr VINCENT, Gregoire; PIMONT, François;
VERLEY, Philippe, 2021, "A note on PAD/LAD estimators implemented in
AMAPVox 1.7", \doi{10.23708/1AJNMP}, DataSuds, V1 } }
}

\section{Contact}{
 \email{contact@amapvox.org}
}

\seealso{
Useful links:
\itemize{
  \item \url{https://github.com/umr-amap/AMAPVox}
  \item Report bugs at \url{https://github.com/umr-amap/AMAPVox/issues}
}

}
\author{
Philippe VERLEY \email{philippe.verley@ird.fr}
}
