% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/AMAPVoxGenerics.R
\docType{methods}
\name{length.VoxelSpace}
\alias{length.VoxelSpace}
\alias{dim.VoxelSpace}
\title{Length of a VoxelSpace}
\usage{
\method{length}{VoxelSpace}(x)

\method{dim}{VoxelSpace}(x)
}
\arguments{
\item{x}{a \code{\link{VoxelSpace-class}} object.}
}
\value{
the number of voxels in the voxel space.

the number of voxels nx, ny, nz of a \code{\link{VoxelSpace-class}} along x, y, z axis.
}
\description{
Get the number of voxels in the voxel space.

Retrieve the dimension of a \code{\link{VoxelSpace-class}}
}
\section{Warning}{

AMAPVox allows to discard empty voxels in the voxel file. In such case
\code{length.VoxelSpace} will return the expected number of voxels as if
none were missing. As a  consequence the number of voxels stored in the
\code{\link{VoxelSpace-class}} object may be inferior to the returned
value, namely \code{nrow(x@voxels) <= length(x)}
}

