% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/AMAPVoxVersionManager.R
\docType{methods}
\name{getLocalVersions}
\alias{getLocalVersions}
\title{List local AMAPVox versions.}
\usage{
getLocalVersions()
}
\value{
a \code{data.frame} with 2 variables: \code{$version} that stores
the version number and \code{$path} the local path of the AMAPVox
directory.
}
\description{
List AMAPVox versions already installed on your computer by
the package. AMAPVox versions are installed in the user-specific data
directory, as specified by \code{\link[rappdirs]{user_data_dir}}.
}
\seealso{
\code{\link{getRemoteVersions}},
\code{\link[rappdirs]{user_data_dir}}
}
