% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/VersionManager.R
\docType{methods}
\name{getRemoteVersions}
\alias{getRemoteVersions}
\title{List remote AMAPVox versions.}
\usage{
getRemoteVersions()
}
\value{
a \code{data.frame} with 2 variables: \verb{$version} that stores
the version number and \verb{$url} the URL of the associated ZIP file.
}
\description{
List AMAPVox versions available for download from AMAPVox
repository \url{https://forge.ird.fr/groups/amap/amapvox/-/packages}
}
\seealso{
\code{\link[=getLocalVersions]{getLocalVersions()}}
}
