% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Ground.R
\docType{methods}
\name{groundLayer}
\alias{groundLayer}
\title{Extract ground layer from voxel space.}
\usage{
groundLayer(vxsp, as.raster = FALSE)
}
\arguments{
\item{vxsp}{a \code{\link{VoxelSpace-class}} object.}

\item{as.raster}{a boolean, whether to return the ground layer as \code{raster} or
a \code{data.table}.}
}
\value{
either a \link[data.table:data.table-class]{data.table::data.table} object with voxel index of
the ground layer or a \link[terra:rast]{terra::rast} raster with 3 layers: k-index,
z-elevation and ground_distance.
}
\description{
Extract ground layer from \code{\link{VoxelSpace-class}} object.
The ground layer is the set of voxels that are just above ground level. The
bottom facet of the voxel must be above ground. Ground layer may be missing
or incomplete for some voxel. A warning is thrown in such case.
}
