% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/AMAPVoxGenerics.R
\docType{methods}
\name{is.VoxelSpace}
\alias{is.VoxelSpace}
\title{Tests for objects of class VoxelSpace}
\usage{
is.VoxelSpace(x)
}
\arguments{
\item{x}{to be tested}
}
\description{
Tests for objects of class VoxelSpace
}
\seealso{
\code{\link{VoxelSpace-class}};
}
