% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/PlantAreaDensity.R
\name{plantAreaDensity}
\alias{plantAreaDensity}
\title{Computes Plant Area Density (PAD)}
\usage{
plantAreaDensity(
  vxsp,
  lad = "spherical",
  angle.name = "angleMean",
  variable.name = c("transmittance", "attenuation_FPL_unbiasedMLE",
    "attenuation_PPL_MLE"),
  pad.max = 5,
  pulse.min = 5,
  ...
)
}
\arguments{
\item{vxsp}{the object of class VoxelSpace}

\item{lad}{the name of the probability density function of the leaf angle
distribution. One of \code{AMAPVox:::leafAngleDistribution}.}

\item{angle.name}{the name of the mean angle variable in the VoxelSpace
object.}

\item{variable.name}{the name of the transmittance/attenuation variables in
the VoxelSpace object. Transmittance variables are expected to start with
"tra" and attenuation variables with "att".}

\item{pad.max}{a float, the maximal PAD value}

\item{pulse.min}{an integer, the minimal number of pulses in a voxel for
computing the PAD. PAD set to NA otherwise.}

\item{...}{additional parameters which will be passed to the leaf angle
distribution functions. Details in \link{computeG}.}
}
\description{
Computes Plant Area Density either from transmittance or
attenuation coefficient estimates.
}
\references{
VINCENT, Gregoire; PIMONT, François; VERLEY, Philippe, 2021,
"A note on PAD/LAD estimators implemented in AMAPVox 1.7",
\url{https://doi.org/10.23708/1AJNMP}, DataSuds, V1
}
\seealso{
\link{computeG}
}
