% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/RemoveButterfly.R
\docType{methods}
\name{removeButterfly}
\alias{removeButterfly}
\title{Remove butterflies from a VoxelSpace object.}
\usage{
removeButterfly(vxsp)
}
\arguments{
\item{vxsp}{a VoxelSpace object}
}
\value{
a cleaned-up voxel space
}
\description{
Remove butterflies from a \code{\link{VoxelSpace-class}} object.

A butterfly refers to a non-empty isolated voxel. Non-empty means that there
is one or more hits recorded in the voxel. Isolated means that voxels in the
\href{https://en.wikipedia.org/wiki/Moore_neighborhood}{Moore neighborhood}
of rank 1 are empty (no hit).

Voxels identified as butterfly are emptied by updating some state variables,
namely:
\itemize{
\item{number of echo set to zero}
\item{intercepted beam surface set to zero (if variable is outputed)}
\item{plant area density set to zero (if variable is outputed)}
\item{transmittance set to one (if variable is outputed)}
\item{any attenuation variable set to zero}
}
}
\examples{
# load a voxel file
vxsp <- readVoxelSpace(system.file("extdata", "tls_sample.vox", package = "AMAPVox"))
# remove butterflies
vxsp <- removeButterfly(vxsp)
}
