% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/AMAPVoxVersionManager.R
\docType{methods}
\name{removeVersion}
\alias{removeVersion}
\title{Remove specific AMAPVox version from local computer.}
\usage{
removeVersion(version)
}
\arguments{
\item{version, }{a valid and existing AMAPVox local version number
(major.minor.build)}
}
\description{
Uninstall specific AMAPVox version from your computer.
}
\examples{
\dontrun{
# uninstall oldest version from your computer
removeVersion(head(getLocalVersions()$version, 1))
}
}
\seealso{
\code{\link{getLocalVersions}}, \code{\link{installVersion}}
}
