% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/AMAPVoxGenerics.R
\docType{methods}
\name{show,VoxelSpace-method}
\alias{show,VoxelSpace-method}
\title{Show \code{\link{VoxelSpace-class}} object.}
\usage{
\S4method{show}{VoxelSpace}(object)
}
\arguments{
\item{object}{a \code{\link{VoxelSpace-class}} object.}
}
\value{
show returns an invisible NULL.
}
\description{
Display the \code{\link{VoxelSpace-class}} object main
characteristics.
}
\examples{
# load a voxel file
vox <- readVoxelSpace(
           system.file("extdata", "als_sample.vox", package = "AMAPVox"))
# show VoxelSpace object
show(vox)
}
