% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Generics.R
\name{tools}
\alias{tools}
\alias{show,VoxelSpace-method}
\alias{print.VoxelSpace}
\alias{length.VoxelSpace}
\alias{dim.VoxelSpace}
\alias{is.VoxelSpace}
\alias{ncol,VoxelSpace-method}
\alias{nrow,VoxelSpace-method}
\alias{nrow.VoxelSpace}
\alias{names.VoxelSpace}
\title{Tools inherited from base R for VoxelSpace object.}
\usage{
\S4method{show}{VoxelSpace}(object)

\method{print}{VoxelSpace}(x, ...)

\method{length}{VoxelSpace}(x)

\method{dim}{VoxelSpace}(x)

is.VoxelSpace(x)

\S4method{ncol}{VoxelSpace}(x)

\S4method{nrow}{VoxelSpace}(x)

nrow.VoxelSpace(x)

\method{names}{VoxelSpace}(x)
}
\arguments{
\item{object}{a \code{\link{VoxelSpace-class}} object.}

\item{x}{a \code{\link{VoxelSpace-class}} object.}

\item{\dots}{further arguments passed to \code{print} function.}
}
\description{
Tools inherited from base R for \code{\link{VoxelSpace-class}}
objects.
}
\section{Note on \code{length.VoxelSpace}}{

AMAPVox allows to discard empty voxels in the voxel file. In such case
\code{length.VoxelSpace} will return the expected number of voxels as if
none were missing. As a  consequence the number of voxels stored in the
\code{\link{VoxelSpace-class}} object may be inferior to the returned
value, namely \code{nrow(x) <= length(x)}
}

