% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/AMAPVox.R
\docType{package}
\name{AMAPVox}
\alias{AMAPVox}
\alias{AMAPVox-package}
\title{AMAPVox package}
\description{
The AMAPVox package provides a a set of R functions for reading, manipulating
and writing voxel spaces. Voxel spaces are read from text-based output files
of the AMAPVox software.
}
\details{
As of version 0.1, available functions are limited and rudimentary, basically
read/write voxel space. With time the package will include functions to
launch full GUI tool from R and a set of useful pre/post-processing functions.
}
\section{References}{

\itemize{
\item{
Research paper firt describing AMAPVox:\cr
Vincent, G., Antin, C., Laurans, M., Heurtebize, J., Durrieu, S.,
Lavalley, C., & Dauzat, J. (2017). Mapping plant area index of tropical
evergreen forest by airborne laser scanning. A cross-validation study
using LAI2200 optical sensor. Remote Sensing of Environment, 198, 254-266.
\url{https://doi.org/10.1016/j.rse.2017.05.034}
}
\item{
Up-to-date description of PAD/LAD estimators implemented in AMAPVox:\cr
VINCENT, Gregoire; PIMONT, François; VERLEY, Philippe, 2021,
"A note on PAD/LAD estimators implemented in AMAPVox 1.7",
\url{https://doi.org/10.23708/1AJNMP}, DataSuds, V1
}
}
}

\section{Contact}{

\email{contact@amapvox.org}
}

\seealso{
Useful links:
\itemize{
  \item \url{https://www.amavox.org}
  \item \url{https://github.com/umr-amap/AMAPVox}
}

}
\author{
Philippe VERLEY \email{philippe.verley@ird.fr}
}
