% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/AMAPVoxVersionManager.R
\docType{methods}
\name{getRemoteVersions}
\alias{getRemoteVersions}
\title{List remote AMAPVox versions.}
\usage{
getRemoteVersions()
}
\value{
a \code{data.frame} with 2 variables: \code{$version} that stores
the version number and \code{$url} the URL of the associated ZIP file.
}
\description{
List AMAPVox versions available for download from page
\url{https://amap-dev.cirad.fr/projects/amapvox/files}
}
\seealso{
\code{\link{getLocalVersions}}
}
