% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/PlotVoxelSpace.R
\docType{methods}
\name{plot,VoxelSpace,ANY-method}
\alias{plot,VoxelSpace,ANY-method}
\title{Plot an object of class VoxelSpace}
\usage{
\S4method{plot}{VoxelSpace,ANY}(
  x,
  y,
  variable.name = "nbSampling",
  palette = "viridis",
  bg.color = "lightgrey",
  width = 640,
  voxel.size = 5,
  unsampled.discard = TRUE,
  empty.discard = TRUE,
  ...
)
}
\arguments{
\item{x}{the object of class VoxelSpace to plot}

\item{y}{Unused (inherited from R base)}

\item{variable.name}{character, the name of the variable to plot}

\item{palette}{character, a valid palette name (one of hcl.pals())}

\item{bg.color}{character, a valid background color name (one of colors())}

\item{width}{numeric, the width of the windows}

\item{voxel.size}{numeric, the size of voxel in pixels}

\item{unsampled.discard}{logical, whether to discard unsampled voxel}

\item{empty.discard}{logical, whether to discard empty voxel (no hit)}

\item{...}{additional parameters which will be passed to
\code{\link[rgl]{plot3d}}.}
}
\description{
plot a \code{\link{VoxelSpace-class}} object.
}
\details{
Plot an object of class VoxelSpace in a 3d device. By default it plots the
sampling intensity but the user can choose any variable available in the
voxel file.
}
\examples{
\dontrun{
# load a voxel file
vxsp <- readVoxelSpace(system.file("extdata", "als_sample.vox", package = "AMAPVox"))
# plot sampling intensity by default
plot(vxsp)
# plot PAD
plot(vxsp, variable.name = "PadBVTotal", palette = "YlOrRd")
}
}
\seealso{
\code{\link[rgl]{plot3d}}
}
