% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ReadVoxelSpace.R
\docType{methods}
\name{readVoxelSpace}
\alias{readVoxelSpace}
\title{Read a voxel file}
\usage{
readVoxelSpace(f)
}
\arguments{
\item{f}{The path of the voxel file.}
}
\description{
read a voxel file and cast it into a
\code{\link{VoxelSpace-class}} object.
}
\seealso{
\code{\link{writeVoxelSpace}}
}
