% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Classes.R
\docType{class}
\name{VoxelSpace-class}
\alias{VoxelSpace-class}
\title{VoxelSpace}
\value{
An object of class VoxelSpace.
}
\description{
Class that holds the state variables of every voxel of the voxel
space in a \code{\link[data.table:data.table-class]{data.table::data.table}} object, plus metadata from the
voxel space header.
}
\section{Slots}{

\describe{
\item{\code{file}}{the path of the voxel file (.vox).}

\item{\code{data}}{the voxels hold in a data.table.}

\item{\code{header}}{a list of parameters associated to this voxel file.}
}}

\seealso{
\code{\link[=readVoxelSpace]{readVoxelSpace()}}
}
