% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/WriteVoxelSpace.R
\docType{methods}
\name{writeVoxelSpace}
\alias{writeVoxelSpace}
\title{Write a voxel file}
\usage{
writeVoxelSpace(voxelSpace, outputFile)
}
\arguments{
\item{voxelSpace}{the object of class VoxelSpace to write}

\item{outputFile}{The path where to write the voxel file.}
}
\description{
write a voxel file out of a \code{\link{VoxelSpace-class}} object.
}
\seealso{
\code{\link{readVoxelSpace}}
}
